Renamed AtomicArray to AtomicMemPool (which is what it is)

+++ YCommon/YContainers/ThreadPool.h
@@ -32,7 +32,7 @@ class ThreadPool {
   }
 
  protected:
-  static int ThreadPoolThread(void* arg);
+  static uintptr_t ThreadPoolThread(void* arg);
 
  private:
   YPlatform::Thread* mThreads;
